test(rpc): relocate dispatch tests under unit_tests and share fixtures via conftest
The session-aware RPC dispatch tests added in #132 were dropped at
tests/utils/rpc/, outside the documented tests/unit_tests/ tree, so CI's
`pytest tests/unit_tests` never collected them. Move them to
tests/unit_tests/rpent/utils/rpc/ and turn the shared helper into conftest
fixtures, matching the placement and fixture conventions in tests/README.md.
